Code review returns: Move method to proper service
